feat(task-26): widen entity primary keys from Int to Long (Room v4 → v5)
All three Room entities (CarEntity, ChargeEventEntity,
CustomLocationEntity) now use `id: Long` instead of `id: Int`, and
ChargeEventEntity.carId becomes `Long`. The change cascades through
DAOs, narrow domain interfaces, all use cases, ViewModel state fields,
Fragment Bundle args, and the navigation safe-args.

DataStore decision:
  ACTIVE_CAR_ID stays an `intPreferencesKey`. Switching the same key
  name to `longPreferencesKey` would silently drop the existing Int
  value at runtime; a single user with thousands of cars over decades
  won't exceed Int.MAX_VALUE. SettingsRepository widens to Long at the
  boundary (Flow<Long> reader + setActiveCarId(id: Long) writer with
  internal id.toInt()) so the rest of the codebase sees the entity-PK
  type consistently.

Migration:
  Room schema bump v4 → v5 with a deliberately **no-op** MIGRATION_4_5.
  SQLite INTEGER columns already store 64-bit signed integers, so
  widening Kotlin Int → Long doesn't change DDL — `diff 5.json 4.json`
  is just the version field. The migration is registered as a tripwire
  so future downgrades trip Room's schema validator instead of
  silently truncating Long values to Int.

  The new schema file `app/schemas/.../5.json` is byte-identical to
  4.json modulo the `version` field, confirming no DDL change.

Backup compatibility:
  BackupData.CURRENT_VERSION 4 → 5, DTO ids widen to Long,
  BackupSerializer.SUPPORTED_VERSIONS = {3, 4, 5}. Gson reads narrower
  JSON Int into wider Kotlin Long fields without coercion, so v3 and
  v4 backups in the wild still restore.

Navigation:
  nav_graph.xml `eventId` argument: app:argType="integer" default="-1"
  → app:argType="long" default="-1L".

Tests:
  - JVM unit tests: ~80 mechanical edits across ~30 files (id literals
    flipped from Int to Long, listOf/setOf assertions L-suffixed,
    factory helper signatures widened). FakeChargeEventWriter,
    FakeCarRepository, FakeSettingsReader/Writer fakes all flip.
  - Instrumented tests: .toInt() casts dropped from
    `dao.insert(...).toInt()` patterns; carId field declarations
    widened.
  - New `migrate_4_to_5_isNoOp_widenIntPksToLong` instrumented test
    runs MIGRATION_3_4 then MIGRATION_4_5 against a v3 fixture and
    asserts existing rows survive untouched with PKs round-tripping
    as 64-bit Long.
  - migrate_1_to_4_validatesSchema renamed to
    migrate_1_to_5_validatesSchema; asserts entity PKs round-trip as
    Long alongside the existing ChargeType.AC and cost-column
    assertions.
  - JVM unit-test count unchanged at 275 (this is a refactor that adds
    no behavior — the migration test is instrumented).

Out of scope (per the design note before starting):
  - TASK-14 (battery degradation tracking — `socBefore` / `socAfter`
    columns, degradation calculator, charts, UI). That ships as a
    separate `MIGRATION_5_6` once Long-PKs are stable.
  - DataStore migration from intPreferencesKey to longPreferencesKey
    for ACTIVE_CAR_ID. Stays Int forever; widening at the repo
    boundary.

All gates green: `:app:assembleDebug`, `:app:assembleRelease` (R8),
`:app:assembleDebugAndroidTest`, `ktlintCheck`, `:app:lint`,
`:app:testDebugUnitTest` (275 cases).

+++ app/src/androidTest/java/org/spsl/evtracker/data/local/dao/CarDaoTest.kt
@@ -40,14 +40,14 @@ class CarDaoTest {
         val rowId = carDao.insert(CarEntity(name = "T", createdAt = 0L))
         assertTrue("rowId must be > 0; got $rowId", rowId > 0)
 
-        val saved = carDao.getById(rowId.toInt())
+        val saved = carDao.getById(rowId)
         assertNotNull(saved)
         assertEquals("T", saved!!.name)
     }
 
     @Test
     fun updateCar_changesPersist() = runBlocking {
-        val rowId = carDao.insert(CarEntity(name = "T", createdAt = 0L)).toInt()
+        val rowId = carDao.insert(CarEntity(name = "T", createdAt = 0L))
         // Re-fetch — the original local still has id = 0 and would silently no-op on update.
         val saved = carDao.getById(rowId)!!
 
@@ -60,8 +60,8 @@ class CarDaoTest {
 
     @Test
     fun deleteCar_removesFromList() = runBlocking {
-        val id1 = carDao.insert(CarEntity(name = "A", createdAt = 0L)).toInt()
-        val id2 = carDao.insert(CarEntity(name = "B", createdAt = 0L)).toInt()
+        val id1 = carDao.insert(CarEntity(name = "A", createdAt = 0L))
+        val id2 = carDao.insert(CarEntity(name = "B", createdAt = 0L))
         // Re-fetch before delete; original local has id = 0.
         carDao.delete(carDao.getById(id1)!!)
