+++ app/src/androidTest/java/org/spsl/evtracker/data/repository/CarRepositoryTest.kt
@@ -38,7 +38,7 @@ class CarRepositoryTest {
 
     @Test
     fun deleteCar_cascadesEvents_throughRepository() = runBlocking {
-        val carId = carRepository.insert(CarEntity(name = "T", createdAt = 0L)).toInt()
+        val carId = carRepository.insert(CarEntity(name = "T", createdAt = 0L))
         repeat(3) {
             chargeEventRepository.insert(
                 ChargeEventEntity(
