+++ app/src/main/java/org/spsl/evtracker/data/local/entity/CarEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "cars")
 data class CarEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0L,
     val name: String,
     val make: String = "",
     val model: String = "",
